# backend/proxmox_api/models/dns_entry_item.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from proxmox_api.models.base_model_ import Model
from proxmox_api import util


class DnsEntryItem(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, entry: str=None):  # noqa: E501
        """DnsEntryItem - a model defined in Swagger

        :param entry: The entry of this DnsEntryItem.  # noqa: E501
        :type entry: str
        """
        self.swagger_types = {
            'entry': str
        }

        self.attribute_map = {
            'entry': 'entry'
        }
        self._entry = entry

    @classmethod
    def from_dict(cls, dikt) -> 'DnsEntryItem':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The DnsEntryItem of this DnsEntryItem.  # noqa: E501
        :rtype: DnsEntryItem
        """
        return util.deserialize_model(dikt, cls)

    @property
    def entry(self) -> str:
        """Gets the entry of this DnsEntryItem.

        dns entry  # noqa: E501

        :return: The entry of this DnsEntryItem.
        :rtype: str
        """
        return self._entry

    @entry.setter
    def entry(self, entry: str):
        """Sets the entry of this DnsEntryItem.

        dns entry  # noqa: E501

        :param entry: The entry of this DnsEntryItem.
        :type entry: str
        """
        if entry is None:
            raise ValueError("Invalid value for `entry`, must not be `None`")  # noqa: E501

        self._entry = entry