# backend/proxmox_api/models/vm_id_item.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from proxmox_api.models.base_model_ import Model
from proxmox_api import util


class VmIdItem(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: float=None):  # noqa: E501
        """VmIdItem - a model defined in Swagger

        :param id: The id of this VmIdItem.  # noqa: E501
        :type id: float
        """
        self.swagger_types = {
            'id': float
        }

        self.attribute_map = {
            'id': 'id'
        }
        self._id = id

    @classmethod
    def from_dict(cls, dikt) -> 'VmIdItem':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The VmIdItem of this VmIdItem.  # noqa: E501
        :rtype: VmIdItem
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> float:
        """Gets the id of this VmIdItem.

        VM id  # noqa: E501

        :return: The id of this VmIdItem.
        :rtype: float
        """
        return self._id

    @id.setter
    def id(self, id: float):
        """Sets the id of this VmIdItem.

        VM id  # noqa: E501

        :param id: The id of this VmIdItem.
        :type id: float
        """
        if id is None:
            raise ValueError("Invalid value for `id`, must not be `None`")  # noqa: E501

        self._id = id